requirements.txt
# Add your project dependencies here
pydantic>=2.0
tqdm
python-dotenv
pandas
scikit-learn
numpy
torch
transformers
sentence-transformers
openai>=1.0.0
